Ssl improvements (#3)
* Use correct variable name to set crypto_method

It's not well documented, but this is the varable that is used (at least
in PHP 7.4 and newer).

See https://github.com/php/php-src/blob/master/ext/openssl/xp_ssl.c#L2574

* Allow the SSL ciphers option to be passed in

The default value for ciphers contains insecure ciphers. Ensure this can
be adjusted by specifying the ssl_ciphers option when creating a socket.